+++ test/java/nginx/clojure/java/AccessHandlerTestSet4NginxJavaRingHandler.java
@@ -1,5 +1,9 @@
 package nginx.clojure.java;
 
+import static nginx.clojure.MiniConstants.DEFAULT_ENCODING;
+import static nginx.clojure.MiniConstants.HEADERS;
+import static nginx.clojure.java.Constants.PHASE_DONE;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Map;
@@ -8,6 +12,7 @@ import java.util.concurrent.Executors;
 
 import javax.xml.bind.DatatypeConverter;
 
+import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
@@ -16,7 +21,6 @@ import org.apache.http.impl.client.HttpClients;
 import nginx.clojure.Configurable;
 import nginx.clojure.NginxHttpServerChannel;
 import nginx.clojure.SuspendExecution;
-import static nginx.clojure.java.Constants.*;
 
 public class AccessHandlerTestSet4NginxJavaRingHandler {
 
@@ -93,7 +97,10 @@ public class AccessHandlerTestSet4NginxJavaRingHandler {
 		public Object[] invoke(Map<String, Object> request) throws SuspendExecution {
 
 			CloseableHttpClient httpclient = HttpClients.createDefault();
+			RequestConfig requestConfig = RequestConfig.custom().setConnectionRequestTimeout(10000).setConnectTimeout(10000)
+					.setSocketTimeout(10000).build();
 			HttpGet httpget = new HttpGet("http://www.apache.org/dist/httpcomponents/httpclient/RELEASE_NOTES-4.3.x.txt");
+			httpget.setConfig(requestConfig);
 			httpget.setHeader("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/37.0.2062.94 Safari/537.36");
 			CloseableHttpResponse response = null;
 			try {